refactor: create edit permission bulk component
